requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiohttp-retry==2.9.1
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
attrs==25.3.0
autobahn==24.4.2
Automat==24.8.1
bidict==0.23.1
boto3==1.37.10
botocore==1.37.10
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
channels==4.2.0
channels_redis==4.2.1
charset-normalizer==3.4.1
colorama==0.4.6
constantly==23.10.4
cryptography==44.0.2
daphne==4.1.2
deprecation==2.1.0
Django==5.1.7
django-channels==0.7.0
django-cors-headers==4.7.0
django-rest-framework==0.1.0
django-storages==1.14.5
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
dnspython==2.7.0
dotenv==0.9.9
eventlet==0.39.1
Faker==37.0.0
filelock==3.17.0
frozenlist==1.5.0
future==1.0.0
gevent==24.11.1
google-ai-generativelanguage==0.6.15
google-api-core==2.24.2
google-api-python-client==2.167.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.3
google-cloud-storage==3.1.0
google-crc32c==1.7.1
google-genai==1.11.0
google-generativeai==0.8.5
google-resumable-media==2.7.2
googleapis-common-protos==1.69.2
gotrue==2.12.0
greenlet==3.1.1
grpcio==1.71.0
grpcio-status==1.71.0
gunicorn==23.0.0
h11==0.14.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.8
httplib2==0.22.0
httpx==0.28.1
hyperframe==6.1.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
iniconfig==2.1.0
jmespath==1.0.1
msgpack==1.1.0
multidict==6.1.0
numpy==2.2.5
oauthlib==3.2.2
packaging==24.2
pgvector==0.4.0
pillow==11.1.0
pluggy==1.5.0
postgrest==1.0.1
propcache==0.3.0
proto-plus==1.26.1
protobuf==5.29.4
psycopg2==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.11.3
pydantic_core==2.33.1
PyJWT==2.9.0
pyOpenSSL==25.0.0
pyparsing==3.2.3
PySocks==1.7.1
pytest==8.3.5
pytest-mock==3.14.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-engineio==4.11.2
python-socketio==5.12.1
pytz==2025.1
realtime==2.4.2
redis==5.2.1
requests==2.32.3
requests-file==2.1.0
requests-oauthlib==2.0.0
rsa==4.9
s3transfer==0.11.4
service-identity==24.2.0
setuptools==76.0.0
simple-websocket==1.1.0
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
storage3==0.11.3
StrEnum==0.4.15
supafunc==0.9.4
tqdm==4.67.1
twilio==9.5.0
Twisted==24.11.0
txaio==23.1.1
typing-inspection==0.4.0
typing_extensions==4.12.2
tzdata==2025.1
uritemplate==4.1.1
urllib3==2.3.0
websockets==14.2
wsproto==1.2.0
yarl==1.18.3
zope.event==5.0
zope.interface==7.2
